Split out castings
