+++ src/WmiScripting/Collections/ArrayCollection.php
@@ -6,6 +6,11 @@ use Illuminate\Support\Collection;
 
 class ArrayCollection extends Collection
 {
+    /**
+     * @param array $items
+     *
+     * @return ArrayCollection
+     */
     public static function collect(array $items = [])
     {
         return new static($items);
